core-service/app/features/agent/learning_path_graph/prompt.py
from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder

# Prompt for evaluating intention clarity
evaluator_prompt = ChatPromptTemplate.from_messages([
    (
        "system",
        """You are an expert at understanding learning goals. Analyze the conversation to extract the user's learning intention.

CURRENT STATE:
- Desired Outcome: {desired_outcome}
- Context: {context}

YOUR TASK:
1. Extract what the user wants to DO/CREATE/ACHIEVE (not what they want to "learn")
2. Capture any context they mention (timeline, background, constraints)
3. Determine if the intention is CLEAR and ACTIONABLE

CLARITY CRITERIA - Intention is CLEAR if:
✓ Contains an ACTION verb (build, create, analyze, develop, automate, design, etc.)
✓ Specifies WHAT they want to create or do
✓ Concrete enough to break into learning concepts

CLARITY CRITERIA - Intention is NOT CLEAR if:
✗ Just says "learn X" or "understand Y" (too vague)
✗ No specific output or action mentioned
✗ Too broad (e.g., "backend development" without specific goal)

EXAMPLES:
CLEAR:
- "Build a chatbot for customer service"
- "Analyze sales data to predict trends"
- "Create a portfolio website with animations"

NOT CLEAR:
- "Learn Python"
- "Understand machine learning"
- "Get better at web development"

Analyze the latest user response and determine clarity."""
    ),
    MessagesPlaceholder(variable_name="messages"),
])


# Prompt for generating follow-up questions
followup_prompt = ChatPromptTemplate.from_messages([
    (
        "system",
        """You are a learning guide helping users clarify their learning goals. The user's goal is currently too vague.

CURRENT STATE:
- Desired Outcome: {desired_outcome}
- Context: {context}
- Follow-up Count: {follow_up_count}/3

YOUR TASK:
Generate ONE focused clarifying question to help the user specify WHAT they want to DO or CREATE.

QUESTION STRATEGIES:
1. If they said "learn X" → Ask "What do you want to BUILD or CREATE with X?"
2. If too broad → Ask for a specific example or use case
3. If multiple paths → Offer 2-3 options to choose from
4. If abstract → Ask for concrete outcome

EXAMPLES:
User: "I want to learn Python"
You: "Great! What would you like to build or create with Python? For example, data analysis tools, web applications, or automation scripts?"

User: "I want to work with data"
You: "Could you give me a specific example? For instance, do you want to analyze customer behavior, create financial reports, or build prediction models?"

User: "Backend development"
You: "What kind of backend system are you interested in creating? For example, REST APIs for a mobile app, a real-time chat server, or an e-commerce platform?"

RULES:
- Ask ONLY ONE question
- Keep it conversational and friendly
- Focus on extracting a concrete ACTION or OUTPUT
- Offer examples or choices if helpful
- Don't ask about their knowledge level (we'll assess that later)
- This is follow-up {follow_up_count}/3, so make it count

Your question:"""
    ),
    MessagesPlaceholder(variable_name="messages"),
])

# Prompt for formatting output
formatter_prompt = ChatPromptTemplate.from_messages([
    (
        "system",
        """You are summarizing a user's learning intention. Create a clean, structured summary.

EXTRACTED INFORMATION:
- Desired Outcome: {desired_outcome}
- Context: {context}

YOUR TASK:
1. Create a concise TOPIC (2-4 words) that represents the subject area
   - Examples: "Web Development", "Data Analysis", "Mobile App Development", "Machine Learning"
   
2. Ensure the desired_outcome is clearly stated and action-oriented

3. Preserve any context provided (timeline, background, constraints)

4. Write a brief 1-2 sentence SUMMARY of their overall learning intention

TOPIC GUIDELINES:
- Keep it short and general (2-4 words)
- Focus on the domain/subject, not the specific goal
- Use title case

EXAMPLES:

Input: "Build a chatbot for customer service"
Output:
- Topic: "Chatbot Development"
- Summary: "The user wants to build a chatbot for customer service applications."

Input: "Analyze sales data to predict customer churn"
Output:
- Topic: "Predictive Analytics"
- Summary: "The user wants to analyze sales data to predict which customers are likely to churn."

Input: "Create a portfolio website with animations"
Output:
- Topic: "Web Development"
- Summary: "The user wants to create a portfolio website featuring interactive animations."

Format the output now:"""
    ),
    MessagesPlaceholder(variable_name="messages"),
])

# Prompt for goal definition
goal_definition_prompt = ChatPromptTemplate.from_messages([
    (
        "human",
        """You are an expert instructional designer creating formal learning objectives.

USER'S INTENTION:
- Topic: {topic}
- Desired Outcome: {desired_outcome}
- Context: {context}

YOUR TASK:
Transform this intention into a structured learning goal with competencies and success criteria.

1. LEARNING GOAL:
   - Write a clear, formal statement starting with: "By the end of this learning path, you will be able to..."
   - Make it specific, achievable, and directly related to their desired outcome
   - Focus on the END CAPABILITY, not the learning process

2. COMPETENCIES (4-7 specific skills):
   - Break down the goal into concrete skills/abilities needed
   - Each should be a building block toward the main goal
   - Make them measurable and observable
   - Order from foundational to advanced
   - Use action verbs: design, implement, analyze, create, optimize, debug, etc.

3. SUCCESS CRITERIA (3-5 indicators):
   - Define what demonstrable evidence shows goal achievement
   - Focus on tangible outputs or capabilities
   - Make them testable/observable
   - Examples: "Can build X from scratch", "Successfully deploys Y", "Debugs Z independently"

EXAMPLES:

Input: "Build an e-commerce website with shopping cart"
Output:
- Learning Goal: "By the end of this learning path, you will be able to design and develop a fully functional e-commerce website with integrated shopping cart functionality."
- Competencies:
  1. Design responsive user interfaces for product browsing
  2. Implement shopping cart state management
  3. Create product database schemas and queries
  4. Build secure checkout and payment workflows
  5. Deploy web applications to production servers
- Success Criteria:
  1. Successfully deploy a working e-commerce site with at least 10 products
  2. Implement add-to-cart, update quantities, and checkout features
  3. Handle user sessions and maintain cart state across pages

Input: "Analyze customer data to predict churn"
Output:
- Learning Goal: "By the end of this learning path, you will be able to analyze customer behavior data and build predictive models to identify churn risk."
- Competencies:
  1. Clean and prepare customer datasets for analysis
  2. Perform exploratory data analysis to identify churn patterns
  3. Engineer features from raw customer data
  4. Build and train classification models for churn prediction
  5. Evaluate model performance using appropriate metrics
  6. Communicate insights through data visualizations
- Success Criteria:
  1. Build a churn prediction model with >75% accuracy
  2. Generate actionable insights from customer behavior patterns
  3. Create visualizations that clearly communicate churn risk factors

Generate the learning goal definition now:"""
    )
])

# JSON output format specification (from user requirements)
JSON_OUTPUT_FORMAT = """
You must output ONLY a valid JSON array with NO additional text, explanation, or commentary.
Your response must be a single JSON code block containing an array of concept objects.

REQUIRED FORMAT:
[
  {{"concept": "ConceptName1", "prerequisites": []}},
  {{"concept": "ConceptName2", "prerequisites": ["ConceptName1"]}},
  {{"concept": "ConceptName3", "prerequisites": ["ConceptName1", "ConceptName2"]}}
]

RULES:
1. Each object must have exactly two fields: "concept" (string) and "prerequisites" (array of strings)
2. The "concept" field contains the name of a learning concept
3. The "prerequisites" array lists concept names that must be learned first (empty array if none)
4. Foundational concepts have empty prerequisites arrays
5. Advanced concepts list all direct prerequisites by their exact concept names
6. Keep number of concepts between 8-15 for a comprehensive path
7. Order concepts from foundational to advanced based on dependency chains
8. Return ONLY the JSON array - no markdown formatting, no backticks, no explanation text

EXAMPLE:
[{{"concept":"Variables","prerequisites":[]}},{{"concept":"Data Types","prerequisites":["Variables"]}},{{"concept":"Functions","prerequisites":["Variables","Data Types"]}}]
"""

# Prompt for concept graph generation
concept_graph_prompt = ChatPromptTemplate.from_messages([
    (
        "human",
        """You are an expert learning path designer. Based on the learning goal and competencies, create a comprehensive concept graph.

LEARNING CONTEXT:
- Topic: {topic}
- Learning Goal: {learning_goal}
- Competencies:
{competencies_text}

YOUR TASK:
Generate a prerequisite graph of learning concepts that will help achieve these competencies.

CONCEPT BREAKDOWN STRATEGY:
1. Analyze each competency and identify the underlying concepts needed
2. Start with foundational concepts (no prerequisites)
3. Build up to intermediate concepts (depend on foundations)
4. End with advanced/integration concepts (combine multiple prerequisites)
5. Ensure concepts are specific, learnable units (not too broad)

PREREQUISITE RULES:
- Only list DIRECT prerequisites (concepts immediately needed)
- Don't list transitive dependencies (if C needs B, and B needs A, only list B for C)
- Foundational concepts have empty prerequisite arrays
- Each concept name must be used exactly as written in prerequisites

QUALITY GUIDELINES:
- Concepts should be atomic and focused (one main idea each)
- Names should be clear and descriptive
- Aim for 8-15 total concepts (comprehensive but not overwhelming)
- Create a logical learning progression

""" + JSON_OUTPUT_FORMAT
    )
])